Added unit test

+++ app/src/main/java/com/larkspur/stockly/Models/IHistoricalPrice.java
@@ -11,8 +11,8 @@ public interface IHistoricalPrice extends Serializable {
 
     public List<Double> getHistoricalPrice();
 
-    public Float getLast24HourChange();
+    public Double getLast24HourChange();
 
-    public Float getLastWeekChange();
+    public Double getLastWeekChange();
 
 }
